Updated plugin to use new Pope module loader through the component registry.
Changed all modules to expose the module manifest allowing for dependency checks.

All existing modules have been ported and now specify an ID. This was mostly necessary to make the module load properly as loading with the new method does not allow a module to load if it doesn't specify a valid ID.

Only basic dependencies have been defined, i.e. to make the plugin "work" but probably many still need tweaks.

The manifest is contained within a comment with the following format:

/***
	{
		Module: module-id,
		Depends: { module-dependency-1, module-dependency-2, module-dependency-3 }
	}
***/

--HG--
branch : module-loader-new