Added auto-update module. At the moment it retrieves the list of products and queries for the license state for all of them. This can be changed in the future (whereby each product "pushes" a request to be put into the list of checked products) and there are a few abstraction in place that should facilitate it.
Also added a new "products" folder, this folder is searched for as a module path added to the registry and any module there is always loaded. This allows for easy inclusion of products and each product module will define its list of dependencies (this could be generated automatically on the server side when the user downloads the plugin zipfile).

--HG--
branch : module_autoupdate