Finished implementation for random/recent gallery image sources.
At the moment only 1 parameter is exposed to pick the amount of images to display, more can be added later.

Originally I was using the same parameter name for the image amount but that created issues as the parameter was set to an array containing the values for the image amount for both random and recent image sources, so switched to using 2 separate parameters.