Made resource loader behave as a singleton utility.
Resources now are properly indexed but unfortunatately because of how they're loaded using gallery skins, all CSS and JS is compressed in a single resource which means the resource loader can't break them down to select which ones need duplication and which one do not.

In order to optimize them and improve performance all the various scripts and styles templates would need to be added as separate resources.