Ensure the plugin loads correctly again.
A few modules, like Attach To Post, use functions in their initialization that requires "init" to be run.

Just turn the Gallery Plugin module class into a real module, with descriptor and "module." name and ensure it's initialized properly after the initi.

We can change this in the future but for now it's just to get things back running.

--HG--
branch : alternative-data-and-storage-layers-using-datamapper
rename : class.photocrati_gallery_plugin.php => module.photocrati_gallery_plugin.php