Add function to retrieve link to the update admin page.
Add the site admin URL to the javascript settings so it can be appended to the request for retrieving or renewing licenses.

Consequently add some logic to append the above parameter to the link the user has to click on. Also make these links open in new tabs/windows.

--HG--
branch : module_autoupdate_admin