Completed porting of flash-based rotator to the new system with rendering fully functional.
Images are now passed to the rotator using the MediaRSS module with the new transient API and transient ID used to store the relevant information.

A small adjustment was required to be able to specify a specific template to use in the MediaRSS URL, this is validated against the 2 available templates for security reasons.

--HG--
branch : gallery-basic-slideshow