Porting of NextGEN legacy code to the new thumbnail/image generation API.
Note that much of the legacy code did not perform multi-stage passes, meaning that when resizing watermarking settings were not applied, same when generating thumbnails no reflections or watermarks were applied etc.

So the new code disables all those features as appropriate to maintain compatibility.

A few other changes in the image generation code were required in order to make this work which is why I'm putting everything in the branch.

--HG--
branch : dynamic-thumbnails