Finished decoupling thumbnail settings from the default nextgen legacy settings.
Now the NextGEN legacy settings are exposed in the global tab under "other options" in the thumbnail settings section.

The basic thumbnail display type now has its own settings for thumbnails which are presented as a local "override" option to ensure compatibility with NextGEN legacy.
By default it will use the same thumbnails as generated in the NextGEN manage galleries tab but you can then override this and force custom settings.