Fix ATP interface disappearing after thumbnail creation. This is due to again as in the past javascript and ready() jQuery handlers being run twice.
I fixed it by ensuring the nggAjax initializers is only called once on the object. Calling it multiple times cause for multiple AJAX actions to be executed and thus running the "complete" event multiple times which called remove_dialog multiple times (and the second time the height of the $window was 0).