Separated size calculation logic from actual rendering/generation for images.
Now the code is more structured, there's a function to retrieve an image clone "result", i.e. information on the image that would be generated if calling generate_image_clone.
Then there are equivalent functions to run through similar processes but with named sizes. These functions simply call an internal function to retrieve a list of parameters based on size names.